// Nebura API/src/typings/discord.d.ts
export interface Precommand {
  /**
   * Name of the precommand.
   * This is the primary identifier used to invoke the command.
   */
  name: string;

  /**
   * Aliases for the precommand.
   * These are alternative names that can also invoke the command.
   */
  aliases?: string[];

  /**
   * Description of the precommand.
   * This provides a brief summary of the command's functionality.
   */
  description: string;

  /**
   * Permissions required by the user to execute the precommand.
   * This is an array of Discord permissions that the user must have.
   */
  permissions: PermissionResolvable[];

  /**
   * Permissions required by the bot to execute the precommand.
   * This is an array of Discord permissions that the bot must have.
   */
  botpermissions: PermissionResolvable[];

  /**
   * Indicates if the precommand is restricted to the bot owner.
   * If `true`, only the bot owner can execute this command.
   */
  owner?: boolean;

  /**
   * Indicates if the precommand is marked as NSFW (Not Safe For Work).
   * If `true`, the command can only be used in NSFW channels.
   */
  nsfw?: boolean;

  /**
   * Cooldown time in seconds for the precommand.
   * This defines the minimum time a user must wait before reusing the command.
   */
  cooldown?: number;

  /**
   * Subcommands associated with the precommand.
   * These are additional commands that extend the functionality of the main command.
   */
  subcommands?: string[];

  /**
   * Usage information for the precommand.
   * This provides guidance on how to use the command.
   */
  usage?: string;

  /**
   * Examples of how to use the precommand.
   * These are sample invocations that demonstrate the command's usage.
   */
  examples?: string[];

  /**
   * Function to execute the precommand.
   *
   * @param client - The main Discord client instance.
   * @param message - The message object that triggered the command.
   * @param args - The arguments provided with the command.
   * @param prefix - The prefix used to invoke the command.
   * @param language - The language code for localization, used to provide responses in the user's language.
   * @param configuration - The bot's configuration object, containing global settings.
   */
  execute: (
    client: MyClient,
    message: Message,
    args: string[],
    prefix: string,
    language: string,
    configuration: typeof config,
  ) => void;
}